Refactor payroll columns to be separate from table logic
