+++ front_end/src/Components/EditPayroll/PayrollNewTable/helpers.js
@@ -3,3 +3,6 @@ export const monthsWithActuals = (previousMonths) => {
     .filter((month) => month.is_actual)
     .map((month) => month.short_name.toLowerCase());
 };
+
+export const totalOfColumn = (data, callback) =>
+  data.reduce((acc, cur) => acc + callback(cur), 0);
